requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
numpy==1.15.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
urllib3==1.23
